Add files via upload
This document is a reference list accompanying the Towards a Model of Intuition Generation project. The list contains academic publications concerning the sampling mechanism introduced in the project. Synthesizing recent insights from cognitive psychology, the project assumes that intuitions are sampled from a heterogeneous and partitioned cognitive structure. Only a small number of intuitive claims can be sampled due to human cognitive limitations. Therefore, the model proposes that the sampling strategy people employ has a computationally efficient default marking the intuitive; that is, sampled claims are probable and valuable.